# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QueueOverviewStatus = typeof QueueOverviewStatus[keyof typeof QueueOverviewStatus];


export const QueueOverviewStatus = {
  open: 'open',
  closed: 'closed',
} as const;

export interface QueueOverview {
  kit: string;
  label: string;
  status: QueueOverviewStatus;
  count: number;
  max: number;
  activeTesters: number;
  channelName: string;
  lastUpdate: string;
}

export interface QueueEntry {
  id: string;
  position: number;
  discordUserId: string;
  username: string;
  ign: string;
  currentTier: string;
  region: string;
  server: string;
  joinedAt: string;
}

export type TicketStatus = typeof TicketStatus[keyof typeof TicketStatus];


export const TicketStatus = {
  open: 'open',
  skipped: 'skipped',
  completed: 'completed',
} as const;

export interface Tester {
  discordUserId: string;
  username: string;
  region: string;
  active: boolean;
}

export interface Ticket {
  id: string;
  kit: string;
  player: QueueEntry;
  tester: Tester;
  status: TicketStatus;
  openedAt: string;
}

export type QueueDetailStatus = typeof QueueDetailStatus[keyof typeof QueueDetailStatus];


export const QueueDetailStatus = {
  open: 'open',
  closed: 'closed',
} as const;

export interface QueueDetail {
  kit: string;
  label: string;
  status: QueueDetailStatus;
  count: number;
  max: number;
  entries: QueueEntry[];
  activeTesters: Tester[];
  currentTicket: Ticket | null;
  lastUpdate: string;
}

export interface QueueActionInput {
  actorId: string;
  actorName: string;
}

export interface WaitlistInput {
  discordUserId: string;
  username: string;
  ign: string;
  currentTier: string;
  region: string;
  server: string;
  kit?: string;
}

export interface QueueActionResponse {
  message: string;
  queue: QueueDetail;
}

export interface VerificationInput {
  discordUserId: string;
  username: string;
  ign: string;
}

export interface VerifiedAccount {
  discordUserId: string;
  username: string;
  ign: string;
  verifiedAt: string;
}

export type ResultInputTier = typeof ResultInputTier[keyof typeof ResultInputTier];


export const ResultInputTier = {
  'N/A': 'N/A',
  lt5: 'lt5',
  ht5: 'ht5',
  lt4: 'lt4',
  ht4: 'ht4',
  lt3: 'lt3',
  ht3: 'ht3',
  lt2: 'lt2',
  ht2: 'ht2',
  lt1: 'lt1',
  ht1: 'ht1',
} as const;

export interface ResultInput {
  playerDiscordUserId: string;
  playerUsername: string;
  ign: string;
  kit: string;
  tier: ResultInputTier;
  testerId: string;
  testerName: string;
  previousTier: string;
  region?: string;
}

export interface TierResult {
  id: string;
  playerUsername: string;
  ign: string;
  kit: string;
  tier: string;
  previousTier: string;
  testerName: string;
  createdAt: string;
  discordMessage?: string;
}

export interface PlayerProfile {
  ign: string;
  username: string;
  bestTier: string;
  tiers: TierResult[];
}

export interface SetupInput {
  guildId: string;
  actorId: string;
}

export interface SetupResult {
  guildId: string;
  createdChannels: string[];
  createdRoles: string[];
  message: string;
}

export type ActivityItemKind = typeof ActivityItemKind[keyof typeof ActivityItemKind];


export const ActivityItemKind = {
  queue: 'queue',
  ticket: 'ticket',
  result: 'result',
  setup: 'setup',
  verification: 'verification',
} as const;

export interface ActivityItem {
  id: string;
  title: string;
  description: string;
  timestamp: string;
  kind: ActivityItemKind;
}