# lib/api-zod/src/generated/types/playerProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TierResult } from './tierResult';

export interface PlayerProfile {
  ign: string;
  username: string;
  bestTier: string;
  tiers: TierResult[];
}