# lib/api-zod/src/generated/types/queueDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QueueDetailStatus } from './queueDetailStatus';
import type { QueueEntry } from './queueEntry';
import type { Tester } from './tester';
import type { Ticket } from './ticket';

export interface QueueDetail {
  kit: string;
  label: string;
  status: QueueDetailStatus;
  count: number;
  max: number;
  entries: QueueEntry[];
  activeTesters: Tester[];
  currentTicket: Ticket | null;
  lastUpdate: Date;
}