# lib/api-zod/src/generated/types/queueOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QueueOverviewStatus } from './queueOverviewStatus';

export interface QueueOverview {
  kit: string;
  label: string;
  status: QueueOverviewStatus;
  count: number;
  max: number;
  activeTesters: number;
  channelName: string;
  lastUpdate: Date;
}