# lib/api-zod/src/generated/types/resultInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResultInputTier } from './resultInputTier';

export interface ResultInput {
  playerDiscordUserId: string;
  playerUsername: string;
  ign: string;
  kit: string;
  tier: ResultInputTier;
  testerId: string;
  testerName: string;
  previousTier: string;
  region?: string;
}