# lib/api-zod/src/generated/types/tester.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Tester {
  discordUserId: string;
  username: string;
  region: string;
  active: boolean;
}