# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QueueEntry } from './queueEntry';
import type { Tester } from './tester';
import type { TicketStatus } from './ticketStatus';

export interface Ticket {
  id: string;
  kit: string;
  player: QueueEntry;
  tester: Tester;
  status: TicketStatus;
  openedAt: Date;
}