# lib/api-zod/src/generated/types/tierResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TierResult {
  id: string;
  playerUsername: string;
  ign: string;
  kit: string;
  tier: string;
  previousTier: string;
  testerName: string;
  createdAt: Date;
  discordMessage?: string;
}