# lib/api-zod/src/generated/types/waitlistInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WaitlistInput {
  discordUserId: string;
  username: string;
  ign: string;
  currentTier: string;
  region: string;
  server: string;
  kit?: string;
}